Move contract utils
