+++ packages/contracts/package.json
@@ -45,7 +45,7 @@
         "@0xproject/dev-utils": "^0.4.2",
         "@0xproject/tslint-config": "^0.4.18",
         "@0xproject/subproviders": "^0.10.1",
-        "@0xproject/sol-cov": "^0.0.10",
+        "@0xproject/sol-cov": "^0.0.11",
         "@types/lodash": "4.14.104",
         "@types/node": "^8.0.53",
         "@types/yargs": "^10.0.0",
